+++ docs/changes.rst
@@ -28,6 +28,11 @@ v0.6.0 (in development)
 - The MPD command ``idle`` is now supported by Mopidy for the following
   subsystems: player, playlist, options, and mixer. (Fixes: :issue:`32`)
 
+- A new frontend :mod:`mopidy.frontends.mpris` have been added. It exposes
+  Mopidy through the `MPRIS interface <http://www.mpris.org/>`_ over D-Bus. In
+  practice, this makes it possible to control Mopidy thorugh the `Ubuntu Sound
+  Menu <https://wiki.ubuntu.com/SoundMenu>`_.
+
 **Changes**
 
 - Replace :attr:`mopidy.backends.base.Backend.uri_handlers` with